pyproject.toml
[tool.poetry]
name = "task-test-liluma-bot"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-telegram-bot = {extras = ["job-queue"], version = "^21.4"}
python-dotenv = "^1.0.1"
matplotlib = "^3.9.2"
openpyxl = "^3.1.5"
pandas = "^2.2.2"
aiosqlite = "^0.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
